finihs atoi

+++ ft_memcmp.c
@@ -6,13 +6,15 @@
 /*   By: pifourni <pifourni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/04 22:49:43 by pifourni          #+#    #+#             */
-/*   Updated: 2025/11/04 22:56:59 by pifourni         ###   ########.fr       */
+/*   Updated: 2025/11/05 22:57:34 by pifourni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
